added-encrypted-logic

+++ content/views.py
@@ -1696,8 +1696,8 @@ class Download_Folder_View(APIView):
         blob_size = blob_client.get_blob_properties().size
         offset = 0
         chunk_size = 1024*1024*100 #100 mb chunk
-        total_chunks = blob_size / chunk_size
-        input_length = chunk_size * total_chunks
+        total_chunks = int(blob_size / chunk_size)
+        input_length = int(chunk_size * total_chunks)
         key='12345'
         while True:
             if offset >= blob_size:
